Fix format.

+++ .pc/fix_format03.patch/grub-core/loader/i386/linux.c
@@ -761,7 +761,7 @@ grub_cmd_linux (grub_command_t cmd __attribute__ ((unused)),
     goto fail;
 
   // verify file to be loaded
-  grub_dprintf("linux", "%s size is: %llu", argv[0], (long long unsigned int)file->size);
+  grub_dprintf("linux", "%s size is: %lu", argv[0], file->size);
   // read_buf = grub_malloc ();
 
   if (grub_file_read (file, &lh, sizeof (lh)) != sizeof (lh))
